+++ src/obs.zsh
@@ -99,7 +99,7 @@ _get_last_vault() {
 
 # Defaults
 VERBOSE=false
-VERSION="4.0.1"
+VERSION="4.1.0"
 
 # --- Helper Functions ---
 
@@ -749,6 +749,11 @@ obs_config() {
     $OBS_PYTHON "$python_cli" config "$@"
 }
 
+obs_link() {
+    local python_cli=$(_get_python_cli) || return 1
+    $OBS_PYTHON "$python_cli" link "$@"
+}
+
 obs_research() {
     local python_cli=$(_get_python_cli) || return 1
     $OBS_PYTHON "$python_cli" research "$@"
@@ -830,6 +835,9 @@ obs() {
             ;;
         "config")
             obs_config "$@"
+            ;;
+        "link")
+            obs_link "$@"
             ;;
         "research")
             obs_research "$@"